Extract HolidayOverview component
