+++ frontend/src/stores/holidayStore.ts
@@ -16,6 +16,10 @@ export class HolidayStore extends DomainStore<Holiday> {
     return this.holidays;
   }
 
+  set entities(entities: Holiday[]) {
+    this.holidays = entities;
+  }
+
   @computed
   get entity(): Holiday | undefined {
     return this.holiday;
@@ -31,20 +35,11 @@ export class HolidayStore extends DomainStore<Holiday> {
   @observable
   holiday?: Holiday;
 
+  protected entityURL = '/holidays/';
+  protected entitiesURL = '/holidays/';
+
   constructor(mainStore: MainStore) {
     super(mainStore);
-  }
-
-  @action
-  protected async doDelete(id: number) {
-    await this.mainStore.api.delete('/holidays/' + id);
-    await this.doFetchAll();
-  }
-
-  @action
-  protected async doFetchAll() {
-    const res = await this.mainStore.api.get<Holiday[]>('/holidays');
-    this.holidays = res.data;
   }
 
   @action
